refactor: group up grpc connectivity-related implementation
